+++ docs/change-log.md
@@ -10,6 +10,7 @@ outline: deep
 * Fixed a bug in the create() method of useVingKind & useVingRecord where props passed in to the function would be overridden by the defaults.
 * Added prepend and append slots to FormSelect component.
 * Fix: int prop didn't generate number form field #92
+* Implemented: have page generator add an s3 drop zone if it detects an s3 relationship #100
 
 ## 2024-04-09
 * Fixed: no type int example in ving schema #82